+++ main.go
@@ -8,6 +8,10 @@ import (
 )
 
 func main() {
+	if err := loadSettings(); err != nil {
+		fmt.Fprintln(os.Stderr, "aviso: config.toml:", err)
+	}
+
 	if len(os.Args) > 1 && os.Args[1] == "ipc" {
 		os.Exit(runIPC(os.Args[2:]))
 	}